Moved library download to app init
