+++ src/App.tsx
@@ -8,6 +8,8 @@ import Workspace from './components/Workspace';
 import { useFileSystemProvider } from './components/providers/FileSystemProvider';
 import WorkspaceProvider from './components/providers/WorkspaceProvider';
 import useImport from './hooks/useImport';
+import useUrlFileWriter from './hooks/useUrlFileWriter';
+import commonLibraries from './etc/libraries.json';
 
 const MyBox = styled(Box)(({ theme }) => ({
   padding: theme.spacing(2),
@@ -24,9 +26,25 @@ export default function App() {
 
   const { error, isLoading } = useImport(importUrl, true);
   const { files } = useFileSystemProvider();
+  const { write, isLoading: isWriting } = useUrlFileWriter();
+  const [isAvailable, setAvailable] = React.useState({});
+
+  React.useEffect(() => {
+    const downloadLibraries = async () => {
+      for (const lib of commonLibraries) {
+        if (lib.downloadOnInit && !isAvailable[lib.url]) {
+          await write(lib.url, (fileName) => {
+            return 'libraries/' + lib.name + fileName.replace(lib.trimFromStartPath, '');
+          });
+          setAvailable((prev) => ({ ...prev, [lib.url]: true }));
+        }
+      }
+    };
+    downloadLibraries();
+  }, []);
 
   // Show a loading indicator during the import.
-  if (isLoading) {
+  if (isLoading || isWriting) {
     return (
       <MyBox>
         <CircularProgress sx={{ marginLeft: '50%' }} />
